main menu, tweaks
-Added simple main menu
-Gameplay scene now fades in slightly when starting
-Fixed being unable to parse ints in yarn variables (they show as singles but work)
-Fixed up compliation directives to allow the codebase to build
-Tweaked some developer objects to only be available in a devBuild or inEditor
-Cleaned up scripts folder
-Deprecated REC_Dialogue. Use REC_NPC
-Made the camera face the oven during the bakery interaction

+++ Assets/_Scripts/REC_NPC.cs
@@ -7,6 +7,7 @@ using Cinemachine;
 using UnityEditor;
 using Yarn.Unity;
 using System.Linq;
+using System;
 
 public class REC_NPC : Receiver
 {
@@ -101,7 +102,7 @@ public class REC_NPC : Receiver
                         }
                         break;
                     case VariableType.Int:
-                        int intResult = GameManager.instance.GetYarnVariable<int>(variable);
+                        Single intResult = GameManager.instance.GetYarnVariable<Single>(variable);
                         Debug.Log("Got yarn int variable " + variable + " as " + intResult.ToString());
                         switch (intComparison)
                         {
